Changed schema for configurable search data
If developers have perfect data, they can define a single JSON
file which lists all of their fields, which the default mapper
can then run thorugh and display. For fields in search which are
harder to get to, developers can define a function to help them
piece together the data they want to display.