Tweaks to ExpiredTokenMiddleware
* Tokens are no longer checked for validity against the Globus API,
instead we check the timestamp of the last login against the token
lifetime, or simply wait for an ExpriedGlobusToken exception. In
either case, the user is redirected to the login page then back to
where they wanted to visit. This also accounts for the variability
of Django Sessions vs Globus Tokens, so each can be set to whatever
and not cause conflicts.