Preview now fetches link/scope from search results.
The addition of 'Service Vars' allows for a structured way to check
search results for pieces of data that change the way the portal
behaves. For Preview, the Globus HTTP link and scope are now fetched
from the search result instead of being configured in settings.py

BREAKING CHANGE: globus_portal_framework.transfer.utils.preview()
now takes a positional arg.

Was:

	preview(user, url, ...)

Now:

	preview(user, url, scope, ...)

Since we're still version 0.1.0 and nobody is likely impacted by this,
(I know our users), I'll go ahead with this change without notice.

The new Service Vars come with a deprecation warning but are still
supported if folks are using old preview settings configs. A couple
portals currently use that so it will be kept until 0.3.0