Fixed expried token redirect bug with valid django session
* It was possible for a user to encounter a redirect bug if they
logged in at a precise time between when their globus tokens expired
and when their django session was still valid. Django would refuse
to log them in due to a valid session, and they would continue to get
redirected back to Django's login because of expired tokens. Users
are now logged out if an exception is raised due to expired tokens.