Fixed demo 'perfdata' field 'remote_file_manifests'
Field was indexed as a dict, when it should have been wrapped in a
list to adhear to the remote file manifest spec. The view was changed
without fixing the search data. This is a temporary fix until we either
fix the perfdata field or migrate to a new test data index