Changed search debugging utilities to be optional
Previously, search debugging was automatically included and revealed
as a public url. This wasn't really a problem for typical uses of the
portal due to searches being done with user tokens. With the
introduction of confidential client searches, this can now reveal
data in the search index that the user should not have permission
to see. In that case, search debugging is now optional and may
not be included in that instance.