Fixed possible bug if user was redirected to custom view after logout
Some URLs depended on the user session, and would cause an error
if the user session was popped. The new URLs depend on the global
context constructued by Globus Portal Framework, which are
guaranteed to be set when a user loads a view as long as that context
is set within settings.py.