Removed 404 and 500 handlers in favor of built-ins
I added custom 404 and 500 error handlers a while back, thinking
those were necessary for custom pages. It turns out they're not,
and simply adding 404.html and 500.html pages are sufficient.

Adding to my prior confusion, I wasn't seeing exceptions and was
seeing 500 errors for 404. It turned out I wasn't adding a logger
for the base 'django', log which was silencing exceptions. Additionally,
a bug in context processing was causing a 500 when I was expecting
404s (and the error was being silenced due to logging).