Fixed date filtering bleeding over into other dates
Previously generated date intervals were very inaccurate, and would
likely end up filtering on the wrong day. For example, searching
on the month of Jan 1st 2019 may return results from Dec and Feb.

Date filtering is now exact for day, month, and year intervals, and
properly clamps the date range filter to the first and last of the
given interval based on the given date.