Fixed date ranges not checking overlapping facets
The previously used mechanism was pretty kludgy, attempting to find
the middle of a given facet in hopes that the selected date str would
match. This mostly worked if the date intervals matched, but failed if
the interval was too precise. It also didn't account for date ranges.

Now date interval facets are checked against the exact intervals using
datetime arithmetic, so any facet that falls within the interval is
considered 'checked'. This works for both ranges and intervals.