Changed Parsing behavior to assume less about future GS changes
It was tempting to zip together content and entry ids into one dict
so that fetching them was an easy dict get operation, instead of
needing to search through entry lists to figure out which entry you
want to use. However, I think this will be detremental in the long
run if Globus Search changes its format anymore.

The new style is simpler, and will properly detect old style results
from the @version on response documents. There is also a new function
called format_result, which can also format results in any format
needed. This is handy for mixing and matching styles, especially if
one particular field is complex and users dont want to change it over
just yet.