Added processing for newer style search results
This removes features from the previous style, adds features, and
changes the format of the values passed to 'fields'.

All features are for list items set on SEARCH_INDEXES.index.fields:

* Passing a single string has been removed.
	* This is now ambiguous with multiple entry ids.
* Passing a two tuple of strings has changed. The second argument
MUST now be a dotted string to a function, a function, or None.
	* Passing a string for a field in content is now ambiguous
	due to multiple entry ids.
* Optionally passing a three tuple item is now supported. The third
item refers to the entry id to use, instead of the entire result.

These changes mandate usage of the new version. The old version
behaves exactly the same way as before, and is completely backwards
compatible.