Upgraded facet/filter handling to properly handle sum/avg faceting
Small refactor to better account for the complexity of facet/filter
handling. Generating a list of active filters moved to its own
function. Fetching a list of active filters for a given facet is
now determined by both the facet type and field, instead of only
the field name. This allows multiple different facet types on the
same field without confusing filters.