fix: Login for Globus SDK v3.3.x
Older DGPF components to enable Preview relied on internal components
of the Globus SDK. The last change in v3.3.0 to make them public
broke compatibility with earlier versions. This change adds one last
check to ensure that v3.3.x+ versions should work as well.

The next version of Globus Portal Framework should remove all of this
when dropping support for earlier versions of the SDK.