fix: Upgrade internals from legacy search result format to 2019-08-27
The changes here are purposefully simplistic to upgrade a very overdue
API result format. The fields system for processing entries is still
not ideal for a few different reasons, and should be made better for
both the ease of configuring fields in DGPF and handling different
kinds of entries. Those will be addressed later. For now, these changes
simply upgrade to the newer API version while preserving backwards
compatibility.