refactor: Remove `result_format_version` from search requests
The result_format_version has been updated to the newer version for
a while, but soon the search service will no longer support specifying
the result_format_version parameter at all.

This is mainly an internal change to DGPF, but does also remove support
for explicitly setting this to older versions. Any portals doing this
are looooong out of date and need to update.