fix: Remove un-used validate_token() function
This function isn't used internally anywhere, and is leftover from
older functionality. Its use was discussed, and it's very much something
we want to discourage for pretty much all use-cases (It's much better to
assume a token is active and deal with it being revoked).

This is an internal function, and more of a refactor, but labeling it
a fix so it's noted in changenotes.