feat: Allow non-Globus users to access public search data
These changes allow for non-Globus users to load public search clients
to view public search data by default, with settings to allow for
raising exceptions to handle special cases for non-globus users when
they should be redirected somewhere else. This is handy if there are
multiple indexes available, and it's desired for a private one to
redirect to a Globus Login if desired (via middleware).