feat: Add Incremental Auth to DGPF
Note: This is a WIP.

There are three main methods of incremental auth:

1. Setting a @scopes_required decorator on views
2. Raising a ScopesRequired() exception from within views
3. Ignoring any Globus SDK exception where auth can be properly handled.

Numbers 1 and 2 are an explicit call to upgrade Auth. 3 is an implicit method
handled in middleware, built on Globus GAREs

Comments are welcome! The code in gclients.py is still pretty rough and will
change a lot.

See also: A test implementation using this PR here:

https://github.com/NickolausDS/DGPF-Incremental-Auth-Example-Repo